+++ tests/rgb.rs
@@ -12,10 +12,8 @@ mod rgb {
     mod integration {
         // TODO: Review after support multi-token transfer
         // mod collectibles;
-        mod collectibles;
-
         mod accept;
-        mod consecutive;
+        mod collectibles;
         mod drain;
         mod dustless;
         mod fungibles;
